% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/separation.R
\name{calculate_between}
\alias{calculate_between}
\title{Calculate between variance in network}
\usage{
calculate_between(graph, set_a, set_b)
}
\arguments{
\item{graph}{The input graph object. It should be either an igraph object or
an edge list matrix/data frame.}

\item{set_a}{The first gene set}

\item{set_b}{The second gene set}
}
\value{
a positive number
}
\description{
No description
}
\keyword{internal}
