% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/separation.R
\name{calculate_within}
\alias{calculate_within}
\title{Calculate within variance}
\usage{
calculate_within(graph, given_set)
}
\arguments{
\item{graph}{The input graph object. It should be either an igraph object or
an edge list matrix / data frame.}

\item{given_set}{The first gene set}
}
\value{
a positive number
}
\description{
No description
}
\keyword{internal}
