% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cronbach.alpha.R
\name{cronbach.alpha}
\alias{cronbach.alpha}
\title{Cronbach's alpha}
\usage{
cronbach.alpha(data)
}
\arguments{
\item{data}{A data frame or matrix contains n subjects * m raters.}
}
\value{
The Cronbach's alpha (unstandardized)
}
\description{
Computes Cronbach's alpha
}
\examples{
library(DTSEA)
library(tibble)

# Load the data
data <- tribble(~x, ~y, ~z, 1, 1, 2, 5, 6, 5, 7, 8, 4, 2, 3, 2, 8, 6, 5)

# Run Cronbach's alpha
cat(cronbach.alpha(data))

}
