% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DTSEA-package.R
\name{DTSEA-package}
\alias{DTSEA-package}
\alias{DTsEA}
\title{The Drug target set enrichment analysis (DTSEA)}
\description{
The DTSEA implements a novel application to GSEA and extends the
adoption of GSEA.
}
\details{
DTSEA
}
