% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_data.R
\name{get_data}
\alias{get_data}
\title{Get extra data}
\usage{
get_data(name)
}
\arguments{
\item{name}{Data name}
}
\value{
A list with the wanted data
}
\description{
Get extra data
}
\examples{
# Do some stuff

data <- get_data("ncbi_list")

}
\keyword{internal}
