% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{random_graph}
\alias{random_graph}
\title{A random graph for the computation of the separation measure}
\format{
An igraph object
}
\usage{
random_graph
}
\description{
The random graph was retrieved from Menche et al (2015).
}
\examples{
library(DTSEA)
data("random_graph", package = "DTSEA")
}
\references{
Menche J, Sharma A, Kitsak M, Ghiassian SD, Vidal M, Loscalzo J, Barabási AL.
Disease networks. Uncovering disease-disease relationships through the
incomplete interactome. Science. 2015 Feb 20;347(6224):1257601. doi:
10.1126/science.1257601. PMID: 25700523; PMCID: PMC4435741.
}
\keyword{datasets}
