% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{example_ppi}
\alias{example_ppi}
\title{An example human gene functional interaction network object}
\format{
An igraph object
}
\usage{
example_ppi
}
\description{
We extracted the gene functional interaction network from
multiple sources with experimental evidence and then integrated them.
}
\examples{
library(DTSEA)
data("example_ppi", package = "DTSEA")
}
\references{
Kanehisa M, Furumichi M, Sato Y, Ishiguro-Watanabe M, Tanabe M. KEGG:
integrating viruses and cellular organisms. Nucleic Acids Res. 2021 Jan 8;49
(D1):D545-D551. doi: 10.1093/nar/gkaa970. PMID: 33125081; PMCID: PMC7779016.
}
\keyword{datasets}
